pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "omni-proof"
version = "0.1.0"
description = "Causal-Multimodal Engine for Creative Performance Attribution"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "dowhy>=0.11",
    "econml>=0.15",
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "pinecone>=5.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "structlog>=24.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "lightgbm>=4.0.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "scikit-learn>=1.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/omni_proof"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
